Split into two scripts
