+++ packages/create-vite-app/package.json
@@ -2,7 +2,20 @@
   "name": "@fastify/create-vite-app",
   "version": "0.0.0",
   "description": "Scaffolding tool for Fastify + Vite applications",
-  "type": "module",
+  "keywords": [
+    "cli",
+    "create-fastify-app",
+    "fastify",
+    "scaffold",
+    "vite"
+  ],
+  "license": "MIT",
+  "author": "Fastify Team",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fastify/fastify-vite.git",
+    "directory": "packages/create-vite-app"
+  },
   "bin": {
     "create-vite-app": "./dist/index.js",
     "fastify-create-vite-app": "./dist/index.js"
@@ -11,38 +24,26 @@
     "dist",
     "templates"
   ],
-  "scripts": {
-    "dev": "node src/index.ts",
-    "build": "node scripts/generate-versions.ts && tsc",
-    "prepublishOnly": "pnpm build",
-    "test": "node --test"
-  },
+  "type": "module",
   "publishConfig": {
     "access": "public"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/fastify/fastify-vite.git",
-    "directory": "packages/create-vite-app"
-  },
-  "keywords": [
-    "fastify",
-    "vite",
-    "create-fastify-app",
-    "scaffold",
-    "cli"
-  ],
-  "author": "Fastify Team",
-  "license": "MIT",
-  "engines": {
-    "node": ">=22.18.0"
+  "scripts": {
+    "build": "pnpm generate-versions && pnpm copy-examples && tsc",
+    "copy-examples": "node scripts/copy-examples.ts",
+    "generate-versions": "node scripts/generate-versions.ts",
+    "prepublishOnly": "pnpm build",
+    "test": "node --test"
   },
   "dependencies": {
     "@clack/prompts": "^0.11.0"
   },
   "devDependencies": {
-    "typescript": "catalog:",
     "@types/node": "catalog:",
+    "typescript": "catalog:",
     "yaml": "^2.8.2"
+  },
+  "engines": {
+    "node": ">=22.18.0"
   }
 }
