+++ packages/create-vite-app/scripts/generate-versions.ts
@@ -1,4 +1,4 @@
-import { cpSync, existsSync, mkdirSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'node:fs'
+import { readFileSync, writeFileSync } from 'node:fs'
 import { join } from 'node:path'
 import { parse } from 'yaml'
 
@@ -14,7 +14,7 @@ const workspace = parse(workspaceYaml) as {
 
 // Read @fastify/vite version from its package.json
 const fastifyVitePkg = JSON.parse(
-  readFileSync(join(rootDir, 'packages', 'fastify-vite', 'package.json'), 'utf-8')
+  readFileSync(join(rootDir, 'packages', 'fastify-vite', 'package.json'), 'utf-8'),
 )
 
 // Build the versions object
@@ -31,40 +31,3 @@ const versions: Record<string, string> = {
 const versionsPath = join(pkgDir, 'src', 'versions.json')
 writeFileSync(versionsPath, JSON.stringify(versions, null, 2) + '\n')
 console.log('Generated src/versions.json with versions:', versions)
-
-// Copy template files from examples/react-vanilla-spa to templates/react-spa
-const SKIP_PATTERNS = ['node_modules', 'dist', '.test.js', '.test.ts']
-
-function shouldSkip(name: string): boolean {
-  return SKIP_PATTERNS.some((pattern) => name.includes(pattern))
-}
-
-function copyTemplate(src: string, dest: string): void {
-  const entries = readdirSync(src, { withFileTypes: true })
-
-  for (const entry of entries) {
-    if (shouldSkip(entry.name)) continue
-
-    const srcPath = join(src, entry.name)
-    const destPath = join(dest, entry.name)
-
-    if (entry.isDirectory()) {
-      mkdirSync(destPath, { recursive: true })
-      copyTemplate(srcPath, destPath)
-    } else {
-      cpSync(srcPath, destPath)
-    }
-  }
-}
-
-const templateSrc = join(rootDir, 'examples', 'react-vanilla-spa')
-const templateDest = join(pkgDir, 'templates', 'react-spa')
-
-// Clean and recreate templates directory
-if (existsSync(join(pkgDir, 'templates'))) {
-  rmSync(join(pkgDir, 'templates'), { recursive: true })
-}
-mkdirSync(templateDest, { recursive: true })
-
-copyTemplate(templateSrc, templateDest)
-console.log('Copied template files to templates/react-spa')
