+++ packages/create-vite-app/tsconfig.json
@@ -8,7 +8,5 @@
     "allowImportingTsExtensions": true,
     "noEmit": false
   },
-  "include": [
-    "src/**/*"
-  ]
+  "include": ["src/**/*"]
 }
